feat: split out utils, rename props for clarity
