+++ src/views/open-api-docs-view/components/open-api-operations/index.tsx
@@ -13,56 +13,11 @@ import type {
 	OperationGroup,
 	OperationProps,
 } from 'views/open-api-docs-view/types'
+// Utils
+import { splitOnCapitalLetters } from './utils/split-on-capital-letters'
+import { addWordBreaks } from './utils/add-word-breaks'
+// Styles
 import s from './open-api-operations.module.css'
-import { ReactNode } from 'react'
-
-/**
- * Split a string on capital letters.
- *
- * Works like `string.split('')`, but splits on capital letters.
- *
- * TODO: split this out, put it somewhere else.
- */
-function splitOnCapitalLetters(str: string): string[] {
-	const lowerCased = str.toLowerCase()
-	const result: string[] = []
-	let buffer = ''
-	for (let i = 0; i < str.length; i++) {
-		/**
-		 * If we have a lowercase character, add it to the buffer.
-		 * If we have an uppercase character, add the buffer to the result,
-		 * and reset the buffer.
-		 */
-		const isLowerCase = str[i] === lowerCased[i]
-		if (isLowerCase) {
-			buffer += str[i]
-		} else {
-			result.push(buffer)
-			buffer = str[i]
-		}
-	}
-	// Append any stray buffer, if it's non-empty
-	if (buffer !== '') {
-		result.push(buffer)
-	}
-	// Return the array of strings
-	return result
-}
-
-/**
- * Given an array of strings,
- * Return the array with `<wbr />` elements inserted between each string.
- *
- * TODO: split this out, put it somewhere else.
- */
-function addWordBreaks(stringArray: string[]): ReactNode[] {
-	return stringArray
-		.map((v: string, idx: number) => {
-			// eslint-disable-next-line react/no-array-index-key
-			return idx === 0 ? [v] : [<wbr key={idx} />, v]
-		})
-		.flat()
-}
 
 /**
  * Render operation items for an OpenApiView.
@@ -85,8 +40,8 @@ export function OpenApiOperations({
 									<OperationHeader
 										className={s.header}
 										slug={operation.slug}
-										headerAriaLabel={operation.operationId}
-										headerSlot={addWordBreaks(
+										headingAriaLabel={operation.operationId}
+										headingSlot={addWordBreaks(
 											splitOnCapitalLetters(operation.operationId)
 										)}
 										method={operation.type}
